requirements.txt
# API相关
longport

# 数据库
mysql-connector-python

# 配置管理
watchdog
jsonschema

# 邮件通知
resend

# 日期时间处理
python-dateutil

# 数据处理
pandas
numpy

# 类型提示
typing-extensions

# 新增依赖
requests
python-dotenv 